+++ hdf5/src/hl/group.rs
@@ -4,12 +4,12 @@ use std::panic;
 use std::ptr::addr_of_mut;
 
 use hdf5_sys::{
-    h5::{hsize_t, H5_index_t, H5_iter_order_t},
+    h5::{H5_index_t, H5_iter_order_t, hsize_t},
     h5d::H5Dopen2,
     h5g::{H5G_info_t, H5Gcreate2, H5Gget_info, H5Gopen2},
     h5l::{
-        H5L_info_t, H5L_iterate_t, H5L_type_t, H5Lcreate_external, H5Lcreate_hard, H5Lcreate_soft,
-        H5Ldelete, H5Lexists, H5Literate, H5Lmove, H5L_SAME_LOC,
+        H5L_SAME_LOC, H5L_info_t, H5L_iterate_t, H5L_type_t, H5Lcreate_external, H5Lcreate_hard,
+        H5Lcreate_soft, H5Ldelete, H5Lexists, H5Literate, H5Lmove,
     },
     h5p::{H5Pcreate, H5Pset_create_intermediate_group},
     h5t::H5T_cset_t,