+++ CHANGELOG.md
@@ -1,10 +1,15 @@
 # Changelog
 
-## hdf5 unreleased
 ## hdf5-types unreleased
 ## hdf5-derive unreleased
-## hdf5-sys unreleased
 ## hdf5-src unreleased
+
+## hdf5-sys unreleased
+- Use edition 2024
+- Add unsafe to all C functions
+
+## hdf5 unreleased
+- Use edition 2024
 
 ## hdf5 0.12.5
 Release date: Jun 04, 2026